Implement arbitrary radix FFT
I took the inspiration from FFTPACK, but wrote it using modern Fortran, so the
code is much simpler. On benchmarks it seems to be as fast or faster as
FFTPACK.

The idea is to factor N into primes (or number 4), and use routines passf2,
passf3, passf4, passf5 and pass() to calculate very fast FFT passes for the
given prime (or 4).